+++ src/logging/logging_internal.hpp
@@ -13,8 +13,9 @@
 inline void log_error(const std::string& logger, const std::string& message) {
     static std::string level = "ERROR";
     // call wilton
-    wilton_logger_log(level.c_str(), level.length(), logger.c_str(), logger.length(),
-            message.c_str(), message.length());
+    wilton_logger_log(level.c_str(), static_cast<int>(level.length()), 
+            logger.c_str(), static_cast<int>(logger.length()),
+            message.c_str(), static_cast<int>(message.length()));
 }
 
 #endif	/* WILTON_LOGGING_LOGGING_INTERNAL_HPP */
